create setting class for clean code

+++ app/src/main/java/de/fau/cs/mad/kwikshop/android/view/SettingFragment.java
@@ -16,13 +16,16 @@ import android.widget.AdapterView;
 import android.widget.ListView;
 import android.widget.Toast;
 
+import java.lang.reflect.Array;
 import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.Locale;
+import java.util.Set;
 
 import de.fau.cs.mad.kwikshop.android.R;
 import de.fau.cs.mad.kwikshop.android.common.AutoCompletionBrandData;
 import de.fau.cs.mad.kwikshop.android.common.AutoCompletionData;
+import de.fau.cs.mad.kwikshop.android.common.Setting;
 import de.fau.cs.mad.kwikshop.android.model.AutoCompletionHelper;
 import de.fau.cs.mad.kwikshop.android.model.DatabaseHelper;
 import de.fau.cs.mad.kwikshop.android.model.SimpleStorageBase;
@@ -38,13 +41,15 @@ public class SettingFragment extends Fragment {
     public static String OPTION_2 = "autocomplete";
     public static String OPTION_3 = "create units";
     public static String OPTION_4 = "create groups";
-    public static CharSequence[] localeSelectionNames = {"Default", "English", "German", "Portuguese"};
-    public static CharSequence[] localeIds = {"default", "en", "de", "pt"};
+    public static CharSequence[] localeSelectionNames = {"Default", "English", "German", "Portuguese", "Russian"};
+    public static CharSequence[] localeIds = {"default", "en", "de", "pt", "ru"};
 
     private View rootView;
     private AlertDialog alert;
     private ArrayList setList;
     private ListView listView;
+    private Context context;
+    private ArrayList<Setting> settingsList;
 
 
 
@@ -63,6 +68,8 @@ public class SettingFragment extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
 
+        context = getActivity().getApplicationContext();
+
         rootView = inflater.inflate(R.layout.fragment_setting, container, false);
         listView = (ListView) rootView.findViewById(android.R.id.list);
 
@@ -71,10 +78,12 @@ public class SettingFragment extends Fragment {
             public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
 
                 // local change
-                changeLocalOption(position);
+                if(settingsList.get(position).getName().equals(getString(R.string.settings_option_2_setlocale)))
+                     changeLocalOption(position);
 
                 // delete history of autocompletion
-                deleteAutoCompletionHistoryOption(position);
+                if(settingsList.get(position).getName().equals(getString(R.string.settings_option_3_deleteHistory)))
+                     deleteAutoCompletionHistoryOption(position);
 
             }
         });
@@ -83,13 +92,36 @@ public class SettingFragment extends Fragment {
         // set title for actionbar
         getActivity().setTitle(R.string.title_activity_settings);
 
-        setList = new ArrayList<String>();
-        setList.add(OPTION_1);
-        setList.add(OPTION_2);
-        setList.add(OPTION_3);
-        setList.add(OPTION_4);
 
-        SettingAdapter objAdapter = new SettingAdapter(getActivity(), R.layout.fragment_setting_row, setList);
+        // list of settings
+        settingsList = new ArrayList<>();
+
+        // Locale Setting
+        Setting setLocale = new Setting(context);
+        setLocale.setName(R.string.settings_option_2_setlocale);
+        setLocale.setCaption(R.string.settings_option_2_desc);
+        settingsList.add(setLocale);
+
+        // Autocompletion deletion
+        Setting setAutoCompletionDeletion = new Setting(context);
+        setAutoCompletionDeletion.setName(R.string.settings_option_3_deleteHistory);
+        setAutoCompletionDeletion.setCaption(R.string.settings_option_3_desc);
+        settingsList.add(setAutoCompletionDeletion);
+
+        // Create units
+        Setting setCreateUnits = new Setting(context);
+        setCreateUnits.setName(R.string.settings_option_3_createUnits);
+        setCreateUnits.setCaption(R.string.settings_option_3_desc2);
+        settingsList.add(setCreateUnits);
+
+        // Create groups
+        Setting setCreateGroups = new Setting(context);
+        setCreateGroups.setName(R.string.settings_option_4_createGroups);
+        setCreateGroups.setCaption(R.string.settings_option_4_desc);
+        settingsList.add(setCreateGroups);
+
+        // Adapter for settings view
+        SettingAdapter objAdapter = new SettingAdapter(getActivity(), R.layout.fragment_setting_row,  settingsList);
         listView.setAdapter(objAdapter);
 
         return rootView;
@@ -99,7 +131,7 @@ public class SettingFragment extends Fragment {
 
     private void changeLocalOption(int position){
 
-        if (setList.get(position).equals(OPTION_1)) {
+
 
             AlertDialog.Builder builder = new AlertDialog.Builder(getActivity());
             //int currentLocaleIdIndex = getActivity().getSharedPreferences(SETTINGS, Context.MODE_PRIVATE).getInt(OPTION_1, 0);
@@ -119,7 +151,7 @@ public class SettingFragment extends Fragment {
 
             alert = builder.create();
             alert.show();
-        }
+
 
     }
 
@@ -151,7 +183,7 @@ public class SettingFragment extends Fragment {
 
 
     public void deleteAutoCompletionHistoryOption(int position) {
-        if (setList.get(position).equals(OPTION_2)) {
+
             AlertDialog.Builder builder = new AlertDialog.Builder(getActivity());
             builder.setTitle(R.string.settings_option_3_check_title);
             builder.setMessage(R.string.settings_option_3_check_text);
@@ -170,7 +202,7 @@ public class SettingFragment extends Fragment {
 
             alert = builder.create();
             alert.show();
-        }
+
     }
 
     private void deleteAutoCompletionHistory() {
